Make the cursor used for the permalink a text cursor.
The default cursor value assigned by Bootstrap for readonly text fields
is "not-allowed", which is rendered as a pointer on Chrome (which isn't
great), but as an "unavailable" icon on Internet Explorer, which is
pretty terrible.

This switches to "text" (i.e. an I-beam), which makes it much more
obvious that the text is selectable anyway.